requirements.txt
click==8.1.7
Jinja2==3.1.2
PyGithub==2.1.1
